+++ src/main/java/core/index/robusttree/RNode.java
@@ -81,7 +81,7 @@ public class RNode {
                 return key.getIntAttribute(dimension);
             case LONG:
                 return key.getLongAttribute(dimension);
-            case FLOAT:
+            case DOUBLE:
                 return key.getFloatAttribute(dimension);
             case DATE:
                 return key.getDateAttribute(dimension);
@@ -98,7 +98,7 @@ public class RNode {
                 return ((Integer) value).compareTo(key.getIntAttribute(dimension));
             case LONG:
                 return ((Long) value).compareTo(key.getLongAttribute(dimension));
-            case FLOAT:
+            case DOUBLE:
                 return ((Float) value).compareTo(key.getFloatAttribute(dimension));
             case DATE:
                 return ((SimpleDate) value).compareTo(key.getDateAttribute(dimension));
