+++ src/main/java/core/utils/TypeUtils.java
@@ -5,7 +5,7 @@ import java.text.SimpleDateFormat;
 import java.util.Comparator;
 
 public class TypeUtils {
-	public static enum TYPE {BOOLEAN,INT,LONG,FLOAT,STRING,DATE,VARCHAR};
+	public static enum TYPE {BOOLEAN,INT,LONG,DOUBLE,STRING,DATE,VARCHAR};
 
 	public static class SimpleDate implements Comparable<SimpleDate>{
 		private int year, month, day;
@@ -133,7 +133,7 @@ public class TypeUtils {
 			if ((Integer)x > (Integer)y) return 1;
 			else if ((Integer)x < (Integer)y) return -1;
 			else return 0;
-		case FLOAT:
+		case DOUBLE:
 			if ((Float)x > (Float)y) return 1;
 			else if ((Float)x < (Float)y) return -1;
 			else return 0;
@@ -170,7 +170,7 @@ public class TypeUtils {
 					return ((Long)o1).compareTo((Long)o2);
 				}
 			};
-		case FLOAT:
+		case DOUBLE:
 			return new Comparator<Object> (){
 				public int compare(Object o1, Object o2) {
 					return ((Float)o1).compareTo((Float)o2);
@@ -205,7 +205,7 @@ public class TypeUtils {
 			return Integer.parseInt(token);
 		case LONG:
 			return Long.parseLong(token);
-		case FLOAT:
+		case DOUBLE:
 			return Float.parseFloat(token);
 		case DATE:
 			byte[] bytes = token.getBytes();
@@ -232,7 +232,7 @@ public class TypeUtils {
 			return (Integer)value - 1;
 		case LONG:
 			return (Long)value - 1;
-		case FLOAT:
+		case DOUBLE:
 			return (Float)value - 0.001;
 		case DATE:
 			SimpleDate d = (SimpleDate)value;
